Update fingerprint generation interface.
With the introduction of UUIDs, we need to be able to generate fingerprints from raw bytes, not just strings. In addition, we need to be able to conditionally set the fingerprint version based on whether or not we're using UUIDs.

Reflects https://github.com/signalapp/libsignal-protocol-java/commit/3662b6d705ae4162ad8b3a242daf35171edbb068

Also NumericFingerprintGeneratorTest.testVectors() passes now